Aufteilung der Robustheitstypen
